Users can now draw trees
Trees are not yet validated and users cannot cut the tree by drawing with their mouse yet. Conceptual information also hasn't been added at the home page, however, users can input their list of implications and then draw the resulting tree in the window, if this tree contains the same literals as the ones in the list, then the official tree is drawn and users can use the toggle options on the right hand side of the screen to alter the truth values of the literals. Creator is currently not used in the project and will definitely be changed if not removed in future versions.